scylla_cdc
Module consumer

A module representing the logic behind consuming the data.

Structs

CDCRow

Represents data from a single row in the CDC log. The metadata can be accessed directly like any other member variable, other columns can be accessed by using the struct methods, e.g. get_value. To get more information about the metadata, see the CDC documentation.

CDCRowSchema

A structure used to map names of columns in the CDC to their indices in an internal vector.

Enums

OperationType

Represents different types of CDC operations. For more information, see the CDC documentation.

Traits

Consumer

Trait used to represent a user-defined callback used for processing read CDC rows. During reading the CDC log, the stream ids are grouped by VNodes. One Consumer is created for each such group when reading of a new generation starts and is destroyed after every row from that generation has been read.

ConsumerFactory

Trait used to represent a factory of Consumer instances. For rules about creating new consumers, please refer to the documentation of [‘Consumer’].