scylla_cdc::consumer
Struct CDCRow

pub struct CDCRow<'schema> {
    pub stream_id: StreamID,
    pub time: Uuid,
    pub batch_seq_no: i32,
    pub end_of_batch: bool,
    pub operation: OperationType,
    pub ttl: Option<i64>,
    /* private fields */
}

Represents data from a single row in the CDC log. The metadata can be accessed directly like any other member variable, other columns can be accessed by using the struct methods, e.g. get_value. To get more information about the metadata, see the CDC documentation.

Fields
stream_id: StreamIDtime: Uuidbatch_seq_no: i32end_of_batch: booloperation: OperationTypettl: Option<i64>
Implementations

impl CDCRow<'_>

pub fn from_row(row: Row, schema: &CDCRowSchema) -> CDCRow<'_>

pub fn get_value(&self, name: &str) -> &Option<CqlValue>

Allows to get a value from the column that corresponds to the logged table. Returns None if the value is null. Panics if the column does not exist in this table. To check if such column exists, use column_exists method.

pub fn take_value(&mut self, name: &str) -> Option<CqlValue>

Allows to take a value from the column that corresponds to the logged table. Leaves None in the corresponding column data. Returns None if the value is null or such column doesn’t exist.

pub fn is_value_deleted(&self, name: &str) -> bool

Allows to get info if a value was deleted in this operation. Panics if the column does not exist in this table or the column is a part of primary key (because these values can’t be deleted). To check if such column exists, use column_deletable method.

pub fn get_deleted_elements(&self, name: &str) -> &[CqlValue]

Allows to get deleted elements from a collection. Returns empty slice if the value is null. Panics if the column does not exist in this table or is not a collection. To check if such column exists, use collection_exists method.

pub fn take_deleted_elements(&mut self, name: &str) -> Vec<CqlValue>

Allows to take deleted elements from a collection. Returns new empty vector if the value is null or such column doesn’t exist. The returned value is always owned. Leaves None in place of taken data.

pub fn column_exists(&self, name: &str) -> bool

pub fn column_deletable(&self, name: &str) -> bool

pub fn collection_exists(&self, name: &str) -> bool

pub fn get_non_cdc_column_names(&self) -> impl Iterator<Item = &str>