scylla_cdc::log_reader
Struct CDCLogReaderBuilder

pub struct CDCLogReaderBuilder { /* private fields */ }

CDCLogReaderBuilder is used to create new CDCLogReader instances.

Example

struct DummyConsumer;

#[async_trait]
impl Consumer for DummyConsumer {
    async fn consume_cdc(&mut self, data: CDCRow<'_>) -> anyhow::Result<()> {
        // ... consume received data ...
        Ok(())
    }
}

struct DummyConsumerFactory;

#[async_trait]
impl ConsumerFactory for DummyConsumerFactory {
    async fn new_consumer(&self) -> Box<dyn Consumer> {
        Box::new(DummyConsumer)
    }
}

let session = SessionBuilder::new().known_node("127.0.0.1:9042").build().await.unwrap();
let (cdc_log_printer, handle): (CDCLogReader, RemoteHandle<anyhow::Result<()>>) = CDCLogReaderBuilder::new()
    .session(Arc::new(session))
    .keyspace("log_reader_ks")
    .table_name("log_reader_table")
    .consumer_factory(Arc::new(DummyConsumerFactory))
    .build()
    .await?;

Implementations

impl CDCLogReaderBuilder

pub fn new() -> CDCLogReaderBuilder

Creates new CDCLogReaderBuilder with default configuration.

Default configuration

start_timestamp: current timestamp

window_size: 60 seconds

safety_interval: 30 seconds

sleep_interval: 10 seconds

end_timestamp: the maximum possible duration - i64::MAX milliseconds

should_load_progress: false

should_save_progress: false,

pause_between_saves: 10 seconds

pub fn session(self, session: Arc<Session>) -> Self

Set session to enable CDCLogReader perform queries. This is a required field for CDCLogReaderBuilder::build() In case it is not set CDCLogReaderBuilder::build() will return error message: failed to create the cdc reader: missing session

pub fn keyspace(self, keyspace: &str) -> Self

Set keyspace of the CDC log table that CDCLogReader instance will read data from. This is a required field for CDCLogReaderBuilder::build() In case it is not set CDCLogReaderBuilder::build() will return error message: failed to create the cdc reader: missing keyspace

pub fn table_name(self, table_name: &str) -> Self

Set table name of the CDC log table that CDCLogReader instance will read data from. This is a required field for CDCLogReaderBuilder::build() In case it is not set CDCLogReaderBuilder::build() will return error message: failed to create the cdc reader: missing table name

pub fn start_timestamp(self, start_timestamp: Duration) -> Self

Set start timestamp from which CDCLogReader instance will start reading from the user specified CDC log table.

pub fn end_timestamp(self, end_timestamp: Duration) -> Self

Set end timestamp to stop CDCLogReader instance reading data from the user specified CDC log table.

pub fn window_size(self, window_size: Duration) -> Self

Set window size for the CDCLogReader instance to read data from the user specified CDC log table window by window. Default window size is 60 seconds.

pub fn safety_interval(self, safety_interval: Duration) -> Self

Set safety interval for the CDCLogReader instance to not read data from the CDC log table that has timestamp later than now. Default safety interval is 30 seconds.

pub fn sleep_interval(self, sleep_interval: Duration) -> Self

Set sleep interval for the CDCLogReader instance to sleep for the specified amount of time after reading and consuming data from all streams from a certain time window. Default sleep interval is 10 seconds.

pub fn consumer_factory( self, consumer_factory: Arc<dyn ConsumerFactory>, ) -> Self

Set consumer factory which will be used in the CDCLogReader instance to create consumers and feed them with data fetched from the CDC log.

pub fn should_load_progress(self, value: bool) -> Self

Set flag indicating whether the CDCLogReader instance should resume progress from set previously checkpoints. Default value is false.

pub fn should_save_progress(self, value: bool) -> Self

Set flag indicating whether the CDCLogReader instance should periodically save progress. Default value is false.

pub fn pause_between_saves(self, pause: Duration) -> Self

Set duration of the pause between creating two consecutive checkpoints. Default value is 10 seconds.

pub fn checkpoint_saver(self, cp_saver: Arc<dyn CDCCheckpointSaver>) -> Self

Set instance of CDCCheckpointSaver trait object to load/save checkpoints.

pub async fn build(self) -> Result<(CDCLogReader, RemoteHandle<Result<()>>)>

Build the CDCLogReader after setting all the options It will fail with an error message if all the required fields are not set. Currently required fields are the following: session, keyspace, table_name, consumer_factory